+++ app/utils/data/vatglasses.ts
@@ -83,7 +83,7 @@ export interface VatglassesSectorProperties {
 let vatglassesActiveAirspaces: VatglassesActiveAirspaces = {};
 let updatedVatglassesPositions: { [countryGroupId: string]: { [vatglassesPositionId: string]: null } } = {};
 
-const ignoredPositions = ['ASIAW', 'ASEAN', 'ASEAS', 'RU-SC', 'RU-CEN', 'RU-WSC', 'RU-ESC', 'RU-WRC', 'RU-ERC', ['56', 'NY']];
+const ignoredPositions = ['ASIAW', 'ASEAN', 'ASEAS', 'RUSC', 'RUCEN', 'RUWSC', 'RUESC', 'RUWRC', 'RUERC', ['56', 'NY']];
 
 function checkIgnoredPosition(id: string, callsign: string) {
     return ignoredPositions.some(x => typeof x === 'string' ? id === x : x[0] === id && callsign.startsWith(x[1]));